Add: Component화, 본문 좋아요 기능 추가
