--- public/data/woojin/recommendData.json
@@ -1,112 +1,0 @@
-[
-  {
-    "id": 1,
-    "name": "Leanne Graham",
-    "userID": "Bret",
-    "email": "Sincere@april.biz",
-    "srcProfileImage": "/images/woojin/UserProfileImages/beach.jpg",
-    "location": "wework1",
-    "srcFeedImage": "/images/woojin/UserFeedImages/fantasy.jpg",
-    "countOfLike": "12110",
-    "content": "버섯은 크게 균근성버섯과 부후성버섯으로 나뉘는데, 균근성버섯은 나무의 뿌리에서 나무와 기생하는 버섯이며 부후성버섯은 죽어가는 나무나 낙엽을 분해하여 양분을 얻는 버섯이다."
-  },
-  {
-    "id": 2,
-    "name": "Ervin Howell",
-    "userID": "Antonette",
-    "email": "Shanna@melissa.tv",
-    "srcProfileImage": "/images/woojin/UserProfileImages/beer.jpg",
-    "location": "wework2",
-    "srcFeedImage": "/images/woojin/UserFeedImages/flamingo.jpg",
-    "countOfLike": "1500",
-    "content": "쁠라밍고"
-  },
-  {
-    "id": 3,
-    "name": "Clementine Bauch",
-    "userID": "Samantha",
-    "email": "Nathan@yesenia.net",
-    "srcProfileImage": "/images/woojin/UserProfileImages/cake.jpg",
-    "location": "wework3",
-    "srcFeedImage": "/images/woojin/UserFeedImages/flower_article.jpg",
-    "countOfLike": "1",
-    "content": "어디서나 매우 흔하게 보이는 다년생초로서 바닥에 딱 붙어서 꽃대 대여섯 개가 쏙 올라온다. 톱니 모양의 잎새와 눈에 확 띄는 노란 꽃이 인상적이다."
-  },
-  {
-    "id": 4,
-    "name": "Patricia Lebsack",
-    "userID": "Karianne",
-    "email": "Julianne.OConner@kory.org",
-    "srcProfileImage": "/images/woojin/UserProfileImages/code.jpg",
-    "location": "wework4",
-    "srcFeedImage": "/images/woojin/UserFeedImages/pineapple.jpg",
-    "countOfLike": "10",
-    "content": "봉리수(鳳梨樹)는 열대 남아메리카 북부 원산 파인애플과의 여러해살이풀이다. "
-  },
-  {
-    "id": 5,
-    "name": "Chelsey Dietrich",
-    "userID": "Kamren",
-    "email": "Lucio_Hettinger@annie.ca",
-    "srcProfileImage": "/images/woojin/UserProfileImages/coffee.jpg",
-    "location": "wework5",
-    "srcFeedImage": "/images/woojin/UserFeedImages/shirt.png",
-    "countOfLike": "101",
-    "content": "티(T) 자 모양의 상의 반팔 속옷을 말한다. 티셔츠의 재질은 100% 면으로 만들어진 것과 폴리에스텔과 섞어서 만들어진 것이 있다."
-  },
-  {
-    "id": 6,
-    "name": "Mrs. Dennis Schulist",
-    "userID": "Leopoldo_Corkery",
-    "email": "Karley_Dach@jasper.info",
-    "srcProfileImage": "/images/woojin/UserProfileImages/flower.jpg",
-    "location": "wework6",
-    "srcFeedImage": "/images/woojin/UserFeedImages/texture.png",
-    "countOfLike": "48501",
-    "content": "빨강(영어: red) 또는 빨간색(빨강색) 또한 적색(赤色) 가시광선을 구성하는 색 중에서 파장이 가장 긴 색이며, 파장의 길이는 약 625~740nm이다."
-  },
-  {
-    "id": 7,
-    "name": "Kurtis Weissnat",
-    "userID": "Elwyn.Skiles",
-    "email": "Telly.Hoeger@billy.biz",
-    "srcProfileImage": "/images/woojin/UserProfileImages/rabbits.jpg",
-    "location": "wework7",
-    "srcFeedImage": "/images/woojin/UserFeedImages/tiger.png",
-    "countOfLike": "400",
-    "content": "식육목 고양이과에 속하는 맹수다. 대한민국의 대표 동물이다. 한국어에서 어린 개체는 개호주라 부른다."
-  },
-  {
-    "id": 8,
-    "name": "Nicholas Runolfsdottir V",
-    "userID": "Maxime_Nienow",
-    "email": "Sherwood@rosamond.me",
-    "srcProfileImage": "/images/woojin/UserProfileImages/dog.jpg",
-    "location": "wework8",
-    "srcFeedImage": "/images/woojin/UserFeedImages/vintage.png",
-    "countOfLike": "490",
-    "content": "hoo.."
-  },
-  {
-    "id": 9,
-    "name": "Glenna Reichert",
-    "userID": "Delphine",
-    "email": "Chaim_McDermott@dana.io",
-    "srcProfileImage": "/images/woojin/UserProfileImages/cat.png",
-    "location": "wework9",
-    "srcFeedImage": "/images/woojin/UserFeedImages/wallpape.jpg",
-    "countOfLike": "4000",
-    "content": "꽃과 다른 식물을 심어서 가꾸는 그릇을 가리킨다. 창가 화분은 창문 바깥에 두는 긴 화분을 가리킨다."
-  },
-  {
-    "id": 10,
-    "name": "Clementina DuBuque",
-    "userID": "Moriah.Stanton",
-    "email": "Rey.Padberg@karina.biz",
-    "srcProfileImage": "/images/woojin/UserProfileImages/head.png",
-    "location": "wework10",
-    "srcFeedImage": "/images/woojin/UserFeedImages/wallpaper2.jpg",
-    "countOfLike": "222200",
-    "content": "은하의 중심은 궁수자리 방향에 위치하고, 이곳이 은하가 가장 밝게 보이는 곳이다. 천구의 적도에 관하여, 은하 북쪽으로 카시오페이아 별자리까지, 남쪽으로 남십자자리 정도까지 지난다. 또 궁수자리에서 서쪽으로 전갈자리, 제단자리, 직각자자리, 남쪽삼각형자리, 컴퍼스자리, 센타우루스자리, 파리자리, 남십자자리, 용골자리, 돛자리, 고물자리, 큰개자리, 외뿔소자리, 오리온자리, 그리고 쌍둥이자리, 황소자리, 마차부자리, 페르세우스자리, 안드로메다자리, 카시오페이아자리, 세페우스자리, 도마뱀자리, 백조자리, 여우자리, 화살자리, 독수리자리, 뱀주인자리, 방패자리를 지난다."
-  }
-]
