+++ src/pages/woojin/variable.scss
@@ -60,3 +60,4 @@ $main_box_border: 1px solid $main_box_border_color;
     font-weight: lighter;
     transform: scale(0.8, 1);
 }
+
