+++ src/styles/common.scss
@@ -1,10 +1,3 @@
 * {
   box-sizing: border-box;
 }
-<<<<<<< HEAD
-
-h1 {
-  font-family: 'Lobster', cursive;
-}
-=======
->>>>>>> master
