--- src/styles/variable.scss
@@ -1,4 +1,0 @@
-<<<<<<< HEAD
-$backgroundColor: transparent;
-=======
->>>>>>> master
